Integrate library routes into regular router
