+++ client/src/components/Libraries/LibrariesList.vue
@@ -63,7 +63,7 @@
                         rows="3" />
 
                     <div v-else-if="row.item.deleted && includeDeleted" class="deleted-item">{{ row.item.name }}</div>
-                    <b-link v-else :to="{ path: `folders/${row.item.root_folder_id}` }">{{ row.item.name }}</b-link>
+                    <b-link v-else :to="{ path: `libraries/folders/${row.item.root_folder_id}` }">{{ row.item.name }}</b-link>
                 </template>
                 <template v-slot:cell(description)="{ item }">
                     <LibraryEditField
