--- client/src/components/Libraries/index.js
@@ -1,8 +1,0 @@
-import router from "./LibraryRouter.js";
-
-export const Libraries = {
-    router,
-    render: (h) => h("router-view"),
-};
-
-export default Libraries;
